requirements.txt
spotipy==2.24.0
ytmusicapi==1.11.0
rich==13.9.4
python-dotenv==1.0.1
requests==2.32.3